refactor: ♻️ extract TimelineCommentItem component (#191)
